Started working on polygon_converters. Some name refactoring.
