+++ pyatlas/doc/atlas_entity.html
@@ -19,7 +19,7 @@
     
 <tr><td bgcolor="#aa55cc"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
 <td width="100%"><table width="100%" summary="list"><tr><td width="25%" valign=top><a href="boundable.html">boundable</a><br>
-</td><td width="25%" valign=top><a href="identifier_conversion.html">identifier_conversion</a><br>
+</td><td width="25%" valign=top><a href="identifier_converters.html">identifier_converters</a><br>
 </td><td width="25%" valign=top></td><td width="25%" valign=top></td></tr></table></td></tr></table><p>
 <table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
 <tr bgcolor="#ee77aa">
