+++ pyatlas/doc/edge.html
@@ -20,7 +20,7 @@
 <tr><td bgcolor="#aa55cc"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
 <td width="100%"><table width="100%" summary="list"><tr><td width="25%" valign=top><a href="atlas_entity.html">atlas_entity</a><br>
 </td><td width="25%" valign=top><a href="entity_type.html">entity_type</a><br>
-</td><td width="25%" valign=top><a href="identifier_conversion.html">identifier_conversion</a><br>
+</td><td width="25%" valign=top><a href="identifier_converters.html">identifier_converters</a><br>
 </td><td width="25%" valign=top><a href="node.html">node</a><br>
 </td></tr></table></td></tr></table><p>
 <table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
