+++ pyatlas/doc/identifier_converters.html
@@ -1,15 +1,15 @@
 
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
-<html><head><title>Python: module identifier_conversion</title>
+<html><head><title>Python: module identifier_converters</title>
 <meta charset="utf-8">
 </head><body bgcolor="#f0f0f8">
 
 <table width="100%" cellspacing=0 cellpadding=2 border=0 summary="heading">
 <tr bgcolor="#7799ee">
 <td valign=bottom>&nbsp;<br>
-<font color="#ffffff" face="helvetica, arial">&nbsp;<br><big><big><strong>identifier_conversion</strong></big></big></font></td
+<font color="#ffffff" face="helvetica, arial">&nbsp;<br><big><big><strong>identifier_converters</strong></big></big></font></td
 ><td align=right valign=bottom
-><font color="#ffffff" face="helvetica, arial"><a href=".">index</a><br><a href="file:/Users/lucascram/Projects/atlas/pyatlas/pyatlas/identifier_conversion.py">/Users/lucascram/Projects/atlas/pyatlas/pyatlas/identifier_conversion.py</a></font></td></tr></table>
+><font color="#ffffff" face="helvetica, arial"><a href=".">index</a><br><a href="file:/Users/lucascram/Projects/atlas/pyatlas/pyatlas/identifier_converters.py">/Users/lucascram/Projects/atlas/pyatlas/pyatlas/identifier_converters.py</a></font></td></tr></table>
     <p><tt>Helpful&nbsp;functions&nbsp;to&nbsp;extract&nbsp;information&nbsp;from&nbsp;Atlas&nbsp;identifiers.</tt></p>
 <p>
 <table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
