+++ pyatlas/doc/polygon_converters.html
@@ -1,0 +1,33 @@
+
+<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
+<html><head><title>Python: module polygon_converters</title>
+<meta charset="utf-8">
+</head><body bgcolor="#f0f0f8">
+
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="heading">
+<tr bgcolor="#7799ee">
+<td valign=bottom>&nbsp;<br>
+<font color="#ffffff" face="helvetica, arial">&nbsp;<br><big><big><strong>polygon_converters</strong></big></big></font></td
+><td align=right valign=bottom
+><font color="#ffffff" face="helvetica, arial"><a href=".">index</a><br><a href="file:/Users/lucascram/Projects/atlas/pyatlas/pyatlas/polygon_converters.py">/Users/lucascram/Projects/atlas/pyatlas/pyatlas/polygon_converters.py</a></font></td></tr></table>
+    <p><tt>Helpful&nbsp;functions&nbsp;to&nbsp;convert&nbsp;between&nbsp;pyatlas&nbsp;types&nbsp;and&nbsp;Shapely&nbsp;polygons.</tt></p>
+<p>
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
+<tr bgcolor="#aa55cc">
+<td colspan=3 valign=bottom>&nbsp;<br>
+<font color="#ffffff" face="helvetica, arial"><big><strong>Modules</strong></big></font></td></tr>
+    
+<tr><td bgcolor="#aa55cc"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
+<td width="100%"><table width="100%" summary="list"><tr><td width="25%" valign=top><a href="shapely.html">shapely</a><br>
+</td><td width="25%" valign=top></td><td width="25%" valign=top></td><td width="25%" valign=top></td></tr></table></td></tr></table><p>
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
+<tr bgcolor="#eeaa77">
+<td colspan=3 valign=bottom>&nbsp;<br>
+<font color="#ffffff" face="helvetica, arial"><big><strong>Functions</strong></big></font></td></tr>
+    
+<tr><td bgcolor="#eeaa77"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
+<td width="100%"><dl><dt><a name="-boundable_to_shapely_box"><strong>boundable_to_shapely_box</strong></a>(boundable)</dt><dd><tt>Convert&nbsp;a&nbsp;pyatlas&nbsp;Boundable&nbsp;type&nbsp;to&nbsp;its&nbsp;Shapely&nbsp;Polygon&nbsp;representation.<br>
+The&nbsp;Shapely&nbsp;Polygon&nbsp;will&nbsp;always&nbsp;be&nbsp;a&nbsp;rectangle.</tt></dd></dl>
+ <dl><dt><a name="-polygon_to_shapely_polygon"><strong>polygon_to_shapely_polygon</strong></a>(polygon)</dt><dd><tt>Convert&nbsp;a&nbsp;pyatlas&nbsp;Polygon&nbsp;to&nbsp;its&nbsp;Shapely&nbsp;Polygon&nbsp;representation.</tt></dd></dl>
+</td></tr></table>
+</body></html>
