+++ pyatlas/pyatlas/spatial_index.py
@@ -1,4 +1,8 @@
-import rtree
+import ctypes
+import shapely.geometry
+from shapely.geos import lgeos as _lgeos
+
+import polygon_converters
 
 
 class SpatialIndex(object):
@@ -26,9 +30,12 @@ class SpatialIndex(object):
 
     def initialize_rtree(self):
         """
-        Initialize an rtree to back this SpatialIndex.
+        Initialize an rtree to back this SpatialIndex. The underlying rtree
+        implementation is immutable, so repeated calls to add() will degrade
+        performance. It is recommended that an rtree-backed SpatialIndex initialize
+        its backing tree upfront.
         """
-        self.tree = rtree.RTree(self.initial_entities)
+        self.tree = _RTree(self.initial_entities)
 
     def initialize_quadtree(self):
         """
@@ -64,3 +71,112 @@ class SpatialIndex(object):
                 results.append(result)
 
         return results
+
+
+class _RTree(object):
+    """
+    A wrapper class for the _CustomSTRtree implementation, which calls into the
+    GEOS C library using machinery from Shapely. _CustomSTRtree is immutable,
+    but this class simulates mutability by maintaining a parallel list of
+    elements, and rebuilding the _CustomSTRtree on each add. For this reason,
+    extensive use of the add() method is not recommended.
+
+    TODO make the _CustomSTRtree implementation mutable
+
+    Note also, this class stores raw AtlasEntity identifiers without any
+    EntityType information. For this reason, users of RTree should avoid adding
+    AtlasEntities with different EntityTypes in the same tree.
+    """
+
+    def __init__(self, initial_entities=None):
+        """
+        Create a new RTree, optionally with an iterable of initial AtlasEntities.
+        """
+        self.contents = []
+        self.tree = None
+
+        if initial_entities is not None:
+            for entity in initial_entities:
+                self.contents.append(entity)
+            self._construct_tree_from_contents()
+
+    def _construct_tree_from_contents(self):
+        """
+        Use the tree's contents list (of AtlasEntities) to rebuild the backing
+        _CustomSTRtree.
+        """
+        contents_shapely_format = [
+            polygon_converters.boundable_to_shapely_box(entity) for entity in self.contents
+        ]
+
+        # pack the arguments in format expected by the _HackSTRtree
+        hacktree_arguments = []
+        for entity, cont in zip(self.contents, contents_shapely_format):
+            hacktree_arguments.append((entity.get_identifier(), cont))
+
+        self.tree = _CustomSTRtree(hacktree_arguments)
+
+    def add(self, entity):
+        """
+        Insert an AtlasEntity into the RTree. The underlying STRtree is
+        immutable, so this method forces a rebuild of the entire tree.
+        """
+        self.contents.append(entity)
+        self._construct_tree_from_contents()
+
+    def get(self, boundable):
+        """
+        Given a Boundable object, get a list of the identifiers of all
+        AtlasEntities within the bounds of the Boundable.
+        """
+        if self.tree is not None:
+            boundable = polygon_converters.boundable_to_shapely_box(boundable)
+            return self.tree.get(boundable)
+        else:
+            raise ValueError('RTree is empty')
+
+
+class _CustomSTRtree(object):
+    """
+    Hack re-implementation of the shapely STRtree. Changes the behaviour of the
+    strtree to allow for insertion of the entity atlas identifier (type long).
+    """
+
+    def __init__(self, items):
+        """
+        Parameter items is a list of tuples, where each tuple lookes like:
+        (entity_id: long, entity_geometry: shapely.geometry.polygon)
+        """
+        self._n_geoms = len(items)
+
+        self._tree_handle = shapely.geos.lgeos.GEOSSTRtree_create(max(4, len(items)))
+        for item in items:
+            _lgeos.GEOSSTRtree_insert(self._tree_handle, item[1]._geom,
+                                      ctypes.py_object(long(item[0])))
+
+        geoms = [item[1] for item in items]
+        self._geoms = list(geoms)
+
+    def __del__(self):
+        if self._tree_handle is not None:
+            _lgeos.GEOSSTRtree_destroy(self._tree_handle)
+            self._tree_handle = None
+
+    def get(self, geom):
+        """
+        Get a list of identifiers of AtlasEntities whose bounds intersect the
+        bounds defined by the geom parameter.
+        """
+        if self._n_geoms == 0:
+            return []
+
+        result = []
+
+        def callback(item, userdata):
+            identifier = ctypes.cast(item, ctypes.py_object).value
+            result.append(identifier)
+
+        _lgeos.GEOSSTRtree_query(self._tree_handle, geom._geom, _lgeos.GEOSQueryCallback(callback),
+                                 None)
+
+        return result
