+++ pyatlas/unit_tests/test_polyline_polygon.py
@@ -7,7 +7,7 @@ from pyatlas.location import Location
 from pyatlas.rectangle import Rectangle
 
 
-class PolyLineTest(unittest.TestCase):
+class PolyLinePolygonTest(unittest.TestCase):
     def setUp(self):
         pass
